Change sound
I am in the process of changing the PlaySound(); function in C to miniaudio since it allows for overlapping sounds easily. However, I have not been able to get it to loop. Please help me if you can.
I also added a log for anything which should only be used if something goes wrong